chore: move dtos to shared project
