+++ Testeger/Shared/DTOs/Enumerations/RoleType.cs
@@ -1,0 +1,15 @@
+﻿using System.ComponentModel;
+
+namespace Testeger.Shared.DTOs.Enumerations;
+
+public enum RoleType
+{
+    [Description("Project Manager")]
+    ProjectManager = 0,
+
+    [Description("Quality Assurance Professional")]
+    QA = 1,
+
+    [Description("Software Developer")]
+    Developer = 2
+}
