+++ Testeger/Shared/DTOs/Requests/CreateProject/CreateProjectRequest.cs
@@ -1,0 +1,7 @@
+﻿namespace Testeger.Shared.DTOs.Requests.CreateProject;
+
+public class CreateProjectRequest
+{
+    public required string Name { get; set; }
+    public required string UserId { get; set; }
+}
