+++ Testeger/Shared/DTOs/Requests/CreateTestCaseResult/CreateTestCaseResultRequest.cs
@@ -1,0 +1,8 @@
+﻿namespace Testeger.Shared.DTOs.Requests.CreateTestCaseResult;
+
+public class CreateTestCaseResultRequest
+{
+    public required string TestCaseId { get; set; }
+    public string? ActualResult { get; set; }
+    public TimeSpan? ElapsedTime { get; set; }
+}
