+++ Testeger/Shared/DTOs/Responses/Project/CreateProjectResponse.cs
@@ -1,0 +1,6 @@
+﻿namespace Testeger.Shared.DTOs.Responses.Project;
+
+public class CreateProjectResponse
+{
+    public required string Id { get; set; }
+}
