+++ Testeger/Shared/DTOs/Responses/TestCase/CreateTestCaseResponse.cs
@@ -1,0 +1,6 @@
+﻿namespace Testeger.Shared.DTOs.Responses.TestCase;
+
+public class CreateTestCaseResponse
+{
+    public required string Id { get; set; }
+}
