+++ Testeger/Shared/DTOs/Responses/TestCaseResult/CreateTestCaseResultResponse.cs
@@ -1,0 +1,6 @@
+﻿namespace Testeger.Shared.DTOs.Responses.TestCaseResult;
+
+public class CreateTestCaseResultResponse
+{
+    public required string Id { get; set; }
+}
