+++ Testeger/Shared/DTOs/Responses/TestRequest/CreateTestRequestResponse.cs
@@ -1,0 +1,6 @@
+﻿namespace Testeger.Shared.DTOs.Responses.TestRequest;
+
+public class CreateTestRequestResponse
+{
+    public required string Id { get; set; }
+}
