+++ Testeger/Shared/Testeger.Shared.csproj
@@ -1,13 +1,15 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <Nullable>enable</Nullable>
     <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
 
   <ItemGroup>
 	<PackageReference Include="Blazored.LocalStorage" Version="4.5.0" />
+	<PackageReference Include="FluentValidation" Version="11.9.2" />
+	<PackageReference Include="Microsoft.AspNetCore.Http.Features" Version="5.0.17" />
 	<PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
     <SupportedPlatform Include="browser" />
   </ItemGroup>
