--- Testeger/Testeger.Application/DTOs/Enumerations/TestCaseStatus.cs
@@ -1,15 +1,0 @@
-﻿using System.ComponentModel;
-
-namespace Testeger.Application.DTOs.Enumerations;
-
-public enum TestCaseStatus
-{
-    [Description("The test case is waiting to be tested.")]
-    Pending = 0,
-
-    [Description("The test case has failed during testing.")]
-    Failed = 1,
-
-    [Description("The test case has been successfully tested and completed.")]
-    Completed = 2,
-}
