--- Testeger/Testeger.Application/DTOs/Requests/Common/PagedRequest.cs
@@ -1,15 +1,0 @@
-﻿namespace Testeger.Application.DTOs.Requests.Common;
-
-public class PagedRequest
-{
-    private const int MaxPageSize = 50;
-    private int _pageSize = 10;
-
-    public int PageNumber { get; set; } = 1;
-
-    public int PageSize
-    {
-        get => _pageSize;
-        set => _pageSize = value > MaxPageSize ? MaxPageSize : value;
-    }
-}
