--- Testeger/Testeger.Application/DTOs/Requests/CreateProject/CreateProjectRequest.cs
@@ -1,7 +1,0 @@
-﻿namespace Testeger.Application.DTOs.Requests.CreateProject;
-
-public class CreateProjectRequest
-{
-    public required string Name { get; set; }
-    public required string UserId { get; set; }
-}
