--- Testeger/Testeger.Application/DTOs/Requests/CreateProject/CreateProjectRequestValidator.cs
@@ -1,17 +1,0 @@
-﻿using FluentValidation;
-
-namespace Testeger.Application.DTOs.Requests.CreateProject;
-
-public class CreateProjectRequestValidator : AbstractValidator<CreateProjectRequest>
-{
-    public CreateProjectRequestValidator()
-    {
-        RuleFor(r => r.Name)
-            .NotEmpty()
-            .MaximumLength(50);
-
-        RuleFor(r => r.UserId)
-            .NotEmpty()
-            .MaximumLength(36);
-    }
-}
