--- Testeger/Testeger.Application/DTOs/Requests/CreateTestCaseResult/CreateTestCaseResultRequest.cs
@@ -1,8 +1,0 @@
-﻿namespace Testeger.Application.DTOs.Requests.CreateTestCaseResult;
-
-public class CreateTestCaseResultRequest
-{
-    public required string TestCaseId { get; set; }
-    public string? ActualResult { get; set; }
-    public TimeSpan? ElapsedTime { get; set; }
-}
