--- Testeger/Testeger.Application/DTOs/Requests/CreateTestCaseResult/CreateTestCaseResultRequestValidator.cs
@@ -1,17 +1,0 @@
-﻿using FluentValidation;
-
-namespace Testeger.Application.DTOs.Requests.CreateTestCaseResult;
-
-public class CreateTestCaseResultRequestValidator : AbstractValidator<CreateTestCaseResultRequest>
-{
-    public CreateTestCaseResultRequestValidator()
-    {
-        RuleFor(r => r.TestCaseId)
-            .NotEmpty()
-            .MaximumLength(36);
-
-        RuleFor(r => r.ActualResult)
-            .MinimumLength(5)
-            .MaximumLength(1500);
-    }
-}
