--- Testeger/Testeger.Application/DTOs/Responses/Project/CreateProjectResponse.cs
@@ -1,6 +1,0 @@
-﻿namespace Testeger.Application.DTOs.Responses.Project;
-
-public class CreateProjectResponse
-{
-    public required string Id { get; set; }
-}
