--- Testeger/Testeger.Application/DTOs/Responses/TestCase/CreateTestCaseResponse.cs
@@ -1,6 +1,0 @@
-﻿namespace Testeger.Application.DTOs.Responses.TestCase;
-
-public class CreateTestCaseResponse
-{
-    public required string Id { get; set; }
-}
