--- Testeger/Testeger.Application/DTOs/Responses/TestCase/TestCaseDetailsResponse.cs
@@ -1,10 +1,0 @@
-﻿namespace Testeger.Application.DTOs.Responses.TestCase;
-
-public class TestCaseDetailsResponse
-{
-    public required string Description { get; set; }
-    public required string PreConditions { get; set; }
-    public required string Steps { get; set; }
-    public required string ExpectedResult { get; set; }
-    public required string Environment { get; set; }
-}
