--- Testeger/Testeger.Application/DTOs/Responses/TestCaseResult/CreateTestCaseResultResponse.cs
@@ -1,6 +1,0 @@
-﻿namespace Testeger.Application.DTOs.Responses.TestCaseResult;
-
-public class CreateTestCaseResultResponse
-{
-    public required string Id { get; set; }
-}
