--- Testeger/Testeger.Application/DTOs/Responses/TestRequest/CreateTestRequestResponse.cs
@@ -1,6 +1,0 @@
-﻿namespace Testeger.Application.DTOs.Responses.TestRequest;
-
-public class CreateTestRequestResponse
-{
-    public required string Id { get; set; }
-}
